% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databox.r
\name{add_metric}
\alias{add_metric}
\title{Add a metric to send to databox}
\usage{
add_metric(dco, key, value, date = NULL, unit = NULL, mattrs = NULL)
}
\arguments{
\item{dco}{databaox client object created by \code{client()} or piped/passed
from other \code{add_metric()} calls}

\item{key}{name of metric (without leading \code{$})}

\item{value}{value of metric (must be numeric)}

\item{date}{(optional) \code{Date} or \code{POSIXct} object which will be
converted to the proper databox API format. NOTE: this auto-converts to UTC for the API}

\item{unit}{(optional) units description for the metric (e.g. \code{USD})}

\item{mattrs}{(optiona) named vector of attribute key/value pairs to add to the metric}
}
\value{
\code{dco} updated with metrics values
}
\description{
Add a metric to send to databox
}
\examples{
\dontrun{
databox::client() \%>\%
  add_metric("sales", 83000, Sys.Date()) \%>\%
  add_metric("sales", 4000, unit="USD") \%>\%
  add_metric("expenses", 87500) \%>\%
  add_metric("sales", 123000, mattrs=c("channel"="in-person", "cheese"="manchego")) \%>\%
  push()
}
}
\references{
\url{https://developers.databox.com/api/}
}

