% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databox.r
\name{client}
\alias{client}
\title{Initialize a databox client and return a databox client object (dco)}
\usage{
client(api_key = databox_api_key())
}
\arguments{
\item{api_key}{databox API key (automatically populated if in the environment)}
}
\value{
databaox client object
}
\description{
Initialize a databox client and return a databox client object (dco)
}
\examples{
\dontrun{
databox::client() \%>\%
  add_metric("sales", 83000, Sys.Date()) \%>\%
  add_metric("sales", 4000, unit="USD") \%>\%
  add_metric("expenses", 87500) \%>\%
  add_metric("sales", 123000,
             mattrs=c("channel"="in-person",
                      "cheese"="manchego")) \%>\%
  push()
}
}

