% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databox-package.R
\docType{package}
\name{databox}
\alias{databox}
\alias{databox-package}
\title{Tools to work with the \href{'databox' 'API'}{https://databox.com/}}
\description{
Tools to work with the \href{'databox' 'API'}{https://databox.com/}
}
\author{
Bob Rudis (@hrbrmstr)
}

