% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.r
\name{databox_api_key}
\alias{databox_api_key}
\title{Get or set DATABOX_API_KEY value}
\usage{
databox_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new dataobx API key for the current environment?}
}
\value{
atomic character vector containing the databox API key
}
\description{
The API wrapper functions in this package all rely on a databox API
key residing in the environment variable \code{DATABOX_API_KEY}. The
easiest way to accomplish this is to set it in the `.Renviron` file in your
home directory.
}
\references{
\url{https://developers.databox.com/api/#get-your-token}
}

