% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databox.r
\name{last_push}
\alias{last_push}
\title{Retrieve the metrics posted in the past push}
\usage{
last_push(api_key = databox_api_key())
}
\arguments{
\item{api_key}{databox API key (automatically populated if in the environment)}
}
\description{
Retrieve the metrics posted in the past push
}
\examples{
\dontrun{
databox::last_push()
}
}
\references{
\url{https://developers.databox.com/api/}
}

