% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databox.r
\name{push}
\alias{push}
\title{Push metrics to databox}
\usage{
push(dco)
}
\arguments{
\item{dco}{databaox client object created by \code{client()} or piped/passed
from other \code{add_metric()} calls}
}
\value{
\code{dco} with an empty \code{data} slot (i.e. you can use this for more
    pipes/calls without fear of duplicating values)
}
\description{
Push metrics to databox
}
\examples{
\dontrun{
databox::client() \%>\%
  add_metric("sales", 83000, Sys.Date()) \%>\%
  add_metric("sales", 4000, unit="USD") \%>\%
  add_metric("expenses", 87500) \%>\%
  add_metric("sales", 123000,
             mattrs=c("channel"="in-person",
                      "cheese"="manchego")) \%>\%
  push()
}
}
\references{
\url{https://developers.databox.com/api/}
}

